WIP moved time representation out of tracker and orion
how we vary time for various computations is a point of volatility, so I'm trying to separate it out. I might need to move it back in Orion if we ever need to control time, which I don't think we do.
Anyways this broke main, so don't pull this yet...