Trying to figure out what's wrong
Some ideas;
expose the calculation of celestial targets in tracker, then regenerate the USNO report for celestial coordinates.
Use this to determine if the error is in abberation, parallax, propermotion(unlikely, these are small effects).
If the error arrises calculating horizon coordinates we know the error is in earth orientation or site location. It can't be refraction because I turned that off
If both calculation are still messed up then the catalog might be bad...

2. Determine the angle between the error and the motion; if it's about zero then we know it is entirely a time scale problem